refactor(api): replace references to lodash utils from lib to src
