+++ api/lib/domain/usecases/get-next-challenge-for-demo.js
@@ -1,5 +1,5 @@
+import { _ } from '../../../src/shared/infrastructure/utils/lodash-utils.js';
 import { logger } from '../../../src/shared/infrastructure/utils/logger.js';
-import { _ } from '../../infrastructure/utils/lodash-utils.js';
 import { AssessmentEndedError } from '../errors.js';
 
 const getNextChallengeForDemo = function ({ assessment, answerRepository, challengeRepository, courseRepository }) {
