--- api/lib/infrastructure/utils/lodash-utils.js
@@ -1,2 +1,0 @@
-// FIXME this file should be deleted after migration
-export * from '../../../src/shared/infrastructure/utils/lodash-utils.js';
